successfully get user infomation and redirect to admin route with user id and token
