+++ package.json
@@ -8,6 +8,7 @@
     "cookie-session": "^1.4.0",
     "cors": "2.8.5",
     "express": "^4.17.1",
+    "express-session": "^1.17.1",
     "jsonwebtoken": "^8.5.1",
     "node-fetch": "^2.6.0",
     "pg": "8.0.3",
@@ -16,6 +17,7 @@
     "react-modal": "3.11.2",
     "react-router-dom": "5.1.2",
     "react-scripts": "3.2.0",
+    "url": "^0.11.0",
     "virtualenv": "0.3.1"
   },
   "scripts": {
